Added an analyze() function for a one-click analysis.
This is more ergonomic than spelling out all of the individual steps, and
should make it easier for the average user. I had hoped that this would be
implemented by another developer in a downstream package that depends on
scrapper, so I wouldn't have to deal with UX issues; but that seems unlikely
due to lack of funding. So, as usual, I'll just do it myself.