Summary quantiles are now stored in a nested data.frame.
This is a bit easier to extract programmatically as users don't have to
re-assemble the column name. In particular, we avoid discrepancies in the
precision of the float->string cast between C++ and R. Now, users can traverse
columns of 'quantiles' in the same order as 'compute.summary.quantiles'; the
column names are also created in R instead of C++ for consistency.