Integrate SE-compatible wrappers from scrapple back into scrapper. (#19)
As it turns out, there wasn't much point in keeping the packages separate
because it's become increasingly unlikely that I'll get any resources for
separate maintainership. So, we might as well merge everything into one package
and keep things simple for the end-user.

We deprecate the analyze() function in favor of analyze.se(). We also direct
readers of the vignette to the analyze.se() function.