Report block IDs with the same type as the input for easier matching.
This only applies to functions that report some kind of per-block result, where
the user might want to extract a result for a specific block without coercing
the block identity to a string to match the row/colnames.

We already do this for the QC functions but we rename the field to 'block.ids',
which is easier to understand than 'block.levels'.